Moved the application overview under the application controller
